Remove cross-domain test from the footer
We had a test for the footer link which opens up a new tab and switches to it.

It'd be nice to avoid needing to test cross domain. As it happens the test doesn't tell us much as this component is designed to accept any link, not specifically the one being tested.

Deleting this scenario also lets us remove some complex helper methods.