Rename engine to citizens_advice_components
This matches the patterns established by other view component engines
like Primer and GOV.UK Components. It includes the organisation name as
a prefix and communicates the function of the gem reasonably well.

Components end up being called as
`CitizensAdviceComponents::PaginationComponent`.

The extra Component suffix is conventional with view components but does
make things a touch verbose. We could remove it but that would go
against convention.