Targeted content: Support number anchors
The HTML spec states that IDs can consist of any character
https://html.spec.whatwg.org/multipage/dom.html#the-id-attribute.
However, querySelector requires selectors that start with numbers to be
escaped.

Previously, if you tried to anchor to an ID starting with a number, e.g.
for paragraph numbers this would throw an error.

For our use case we can normalise the hash and use getElementById
instead.

Also wrapping the hash behaviour in a try/catch as this is an
enhancement and it's fine to fail silently if this code fails.